Rework into a Docker container action (#9)
* Rework into a Docker container action

Rather than handling input validation inside the composite action with the
difficulty of dealing with an unknown set of environment variables, rework it
into a Docker container action. The entrypoint.sh in Bash takes care of input
validation and execution behavior.

* Fix entrypoint execution mode

* Fix syntax error

* Really make evaluation conditional

* Copy recursively

* Fix bad substitution

* Use relative path

The workdir is set to /github/workspace so everything should just work out of
the box.

* Temporary debugging

* Ensure the results directory is available

* Correct the reporting step to its original

Some errors were introduced by rewriting rather than doing a copy-paste.

* Correctly close if statement

* Fix if statement test syntax

* Correct report input command

* Update example npm scan

Adopt the new functionality from the Docker container Action in the example
workflow.

* Remove config-dir input from the license scan action

* Correct example with NPM_CONFIG_USERCONFIG environment variable

The NODE_CONFIG_USERCONFIG was falsely used and this is corrected to the
NPM_CONFIG_USERCONFIG variable.